Implement update-namespace hook in jbossas-7 cart
The standalone.xml configuration file sources environment vars
from OPENSHIFT_JBOSS_CLUSTER which contains the entire universe
of cluster participant hostnames. Upon namespace alter, the
file needs updated and the JBoss server restarted in order
to reform the cluster with the new hostnames. In addition to
all this, the cluster configuration must have a lengthy timeout
to ensure DNS for the new namespace has propagated.

To ensure the cluster survives the domain alter, we:

- Rewrite the OPENSHIFT_JBOSS_CLUSTER file in the hook

- Restart the application gear in the hook

- Specify a 30 second timeout in the cluster configuration
  within standalone.xml

This resolves bugs:

- https://bugzilla.redhat.com/show_bug.cgi?id=811671