Disable restart of JBoss app on namespace alter
The application restart which is invoked during the update namespace
hook processing for JBoss carts is suspected to cause intermittent
timeouts in other processes, leaving the namespace alter incomplete.

To be absolutely sure the namespace alter is solid, replace the
restart call with a client message advising the user to restart their
application manually following the alter.