Add php-process to the PHP cartridge.
Some apps depend on being able to spawn daemons, and in PHP this
is done by requiring the POSIX extension (which is contained in
php-process on RHEL). One such example is Facebook's Phabricator
code review apps, which depends on the ability to launch daemons
in order for SCM repositories to be tracked properly.