Set up .ssh directory during gear move
Ensure that observe_setup_var_lib_dir is invoked against the .ssh directory
of the cartridge during a gear move. The call is made explicitly in the
configure hook, and must be invoked in the same way during a gear move
operation.

Clean up some old calls to observe_* functions which are inapplicable to
a catridge such as HAProxy which does not own the APP_ROOT.