Disable Jenkins built-in SSH server by default
The built-in Jenkins SSH server does not provide any means to configure
the interface on which it binds.  This makes the Jenkins server volatile
depending on the restrictions of the node security implementation.

Disable the server by default to ensure consistent startup success.  The
feature can be re-enabled by users on-demand if their environment
is permissive of the binding.