Don't write to pearrc during build
The build is executed in the context of the gear user. The pearrc
file is not writable by the gear user. The build does not modify
the PHP ini setting in the pearrc, and so there's no reason to
attempt to update pearrc to re-set this configuration item.
Attempting to do so will fail, as pearrc is not writable in this
context.

Remove the pearrc PHP ini config update during the build process.

Resolve this bug: https://bugzilla.redhat.com/show_bug.cgi?id=861550