Sanity check CLOUD_USER value
The OpenShift::UnixUser class interpolates the value of CLOUD_USER in
/etc/openshift/node.conf into the name of a shell environment variable.
Therefore, if CLOUD_USER contains symbols that may not appear in a shell
variable name, we will get mysterious errors.

This patch adds a check in the initializer for OpenShift::UnixUser to
verify that CLOUD_USER contains only valid symbols.