AuthService fix for auth plug-ins
Before this patch, the auth plug-ins are attempting to re-open the
module OpenShift::AuthService, which conflicts with a declaration of
OpenShift::AuthService as a class in the controller.

After this patch, the auth plug-ins instead re-open their respective
subclasses of the OpenShift::AuthService class: KerberosAuthService,
MongoAuthService, and RemoteUserAuthService.

The intent is (and was before) that, for example, KerberosAuthService
can be assigned as OpenShift::AuthService.provider.  The plug-ins should
not be modifying OpenShift::AuthService directly; they should be
self-contained in their subclasses.  The erroneous re-opening was added
in commit 353102deaad817ff7a05ee0b70c02bbc55659b8e.