Make stop function work when called prior to any app start
If the stop function is called prior to the app ever having been
started, no PID file exists, and the stop function will halt
the entire script execution with an error while trying to cat
the non-existant PID file. This PID file check is already performed
by the isrunning function, and the lack of a PID file is a valid
state (even according to isrunning); so, the stop function can't
assume the file exists.

Remove the redundant check and fix the stop function when invoked
in this context.