Decouple hook subprocess streams from the parent process
The resolve_application_dependencies is used to call Git hooks asynchronously
during cartridge configure when a template is provided. The async process can
outlive the invoking process (e.g. the MCollective agent). In such a case,
it's possible the subprocess will no longer have a stdout/stderr consumer, and
can hang indefinitely writing to the stream buffers. This output is not
presented to any client, or logged anywhere. For now, redirect the subprocess
streams to /dev/null pending a proper async solution at a higher tier which
would allow these calls to be synchronous.