oo-accept-broker: drop sysctl checks
In oo-accept-broker, we have the check_kernel_settings function which
checks some sysctl settings.  However, we only change these settings on
the node, so these checks do not belong in oo-accept-broker.

The invocation of check_kernel_settings is commented out, so the checks
were never run.  This commit deletes both the commented-out invocation
and the definition of check_kernel_settings.