Replace expose/show/conceal-port hooks with Endpoints
As part of the ongoing effort to implement the new cartridge API, remove the
expose-port, show-port, and conceal-port hooks from the cartridges, and
move their functionality up into the platform via the node library.

Proxy ports are now defined by cart authors via an Endpoints element in the
cartridge manifest YAML, and the platform functionality is implemented
internally within the node library as opposed to being handled ad-hoc in
cartridge hooks.

The show-port and conceal-port hooks are now no-op stubs in the agent
and broker proxy.