Bug 887353: removing a cartridge leaves info/ dir
This commit adds each cartridge's %{cartridgedir}/ directory as well as its
%{cartridgedir}/info/ directory to the %files stanza of the cartridge so
that RPM recognises these directories as belonging to the cartridge and
removes them when the RPMs are removed.

Following the example of the python-2.6 cartridge, which already has these
directories in its %files stanza, this commit uses %dir for the
%{cartridgedir}/ and %{cartridgedir}/info/.  We do need the default
permissions (0755) on these two directories so that the MCollective agent
can read the info/ directory.

This commit fixes Bug 887353.