setup the namespace configuration oo-setup-node
By default, on Fedora 18 and surely others, polyinstanciation is
blocked by selinux. So someone installing pam_openshift on a stock
fedora would not be able to connect as regular user, which would
confuse most users. So in order to respect a principal of least
surprise, and to make sure that the bolean is enabled, the configuration
in /etc/security/namespace.d/ is moved in oo-setup-node ( and cleaned, since
neither gdm nor liveuser belong there, this should be a custom configuration
for the livecd )