disable viminfo creation on node
each time I connect by ssh to a gear and run vim, vim try to
create a .viminfo file in my home, where I do not have the right to create
anything. So either we create a file with proper permissions or
a different location, or we disable .viminfo. I choose the last one.