<threaddump> fix bug 923405 so ruby thread dump reports useful log file
https://bugzilla.redhat.com/show_bug.cgi?id=923405 thread dumps report the wrong log file name due to timezone

When ruby does a thread dump it goes in the httpd server log. This is
rotated via rotatelogs and the name includes the timezone in it. Due to
various complications this does not always match the TZ from `date`.
Rather than try to give the exact file name, we now just give a glob
that is good enough.