Bug 949844: Add support for threaddump command in v2 Ruby cartridge.
In the v1 cartridge, we used a custom thread dumper library to respond to QUIT signal and dump thread information to the log file. (See, e.g., https://github.com/openshift/origin-server/blob/openshift-origin-cartridge-ruby-1.9-scl-1.7.4-1/cartridges/openshift-origin-cartridge-ruby-1.9-scl/template/thread-dumper.rb)
This is not a standard Ruby library, and does not work with users' applications in general.

Instead, in the v2 cartridge, we rely on Passenger's handling of ABRT signal (http://www.modrails.com/documentation/Users%20guide%20Apache.html#debugging_frozen) to dump thread state to the log file. Do note that this will terminate the current application; this is a divergence from the previous behavior, and the users should be duly informed.