Bug 949844
Fix the logic in PID detection. If $PID is empty, warn and abort operation.
At the moment PID detection relies on the presence of a process with "Rack:" in it. This may fail if the application had just been started.

Update README on threaddummp command on the v2 Ruby cartridge