Bug 957045 - fixing oo-accept-systems for v2 cartridges
* The rest-client dependency needed to be loaded before the rails environment
because Bundler would prevent it from loading.

* Cartridge lookup is completely different with v2.  I decided to use
oo-cartridge-list for this in openshift-facts.rb since it supports both v1 and
v2.  The complication with this is that previously oo-cartridge-list would
access the platform logs.  Since openshift_facts.rb is ultimately called by a
cronjob running as openshift_cron_t this would be denied and munge the
oo-cartridge-list output.  As with other admin cli scripts we decided to stub
out the logging class to prevent this from happening.

* the embed_cart_list fact did not appear to be used so I removed it.