<cache> bug 957818 - clear-cache script, oo-cart-version uses it
https://bugzilla.redhat.com/show_bug.cgi?id=957818
oo-cart-version was making some assumptions that didn't work across
projects or usage. It should not assume it is run on the broker, and
it should not assume development mode.

Dealing with clearing cache generically across projects is annoying
enough that I added a script to do it. oo-admin-broker-cache is aware
of SCL and loads the broker rails environment (which is of course aware
of dev/prod mode) to clear the cache.
We should never again need to explain this to people who just want to
clear their cache.

oo-cart-version now can just call this script, if it is even run on a
broker. If it isn't, which is the normal case, it can give the user a
simple command to run themselves.