<console> Bug 959162 - Fix display issues
https://bugzilla.redhat.com/show_bug.cgi?id=959162

Authorizations and sshkeys on the Settings tab were being displayed on a
white background making it hard to distinguish the text, this updates
them to put them in a well similar to how the domains were already being
handled and fixes the styling for stried tables.

Fixing config file logic for openshift console in dev and test configurations
fixing up environment configs and non-prod console config location

removing help_helper.rb from openshift-console, since it exists in
console and doesn't need to be duplicated.