BZ963827 - Logs generated by oo-admin-ctl-gears should include timestamps
The code was actually logging in two different places. The NodeLogger
was already logging to /var/log/openshift/node/platform.log, which
includes the timestamps. The /var/log/openshift-gears-async-start.log
existed before the platform.log and was just a redirection of stdout,
which did not include timestamps. The openshift-gears-async-start.log
is removed with this commit.