<oo-admin-ctl-cgroups> Bug 964205 - add "repair" command
This commit adds a new command to oo-admin-ctl-cgroups called
"repair". This command identifies users which lack a cgroup and
creates a cgroup for them. It can be called with an optional username
argument (or a quoted space-separated list of usernames) to restrict
the action to a specific subset.

The "restartall" command remains functionally unchanged - this will be
addressed in the next commit