Bug 968834: fix `service openshift-console reload`
The openshift-console service fails to complete the reload command
successfully. This appears to be related to how the $OPTIONS variable
is expanded in the configuration syntax check conditional.

This commit is basically the openshift-console version of bugs: 902630
and 902633. Likewise, this commit fixes the unrelated issue where the
reload action does not print anything on success. Finally, this makes
some indentation uniform with surrounding statements and correctly
redirects pushd/popd output to /dev/null.