Add cartridge version override support to CartridgeRepository
While we await higher tiers of the stack to support cartridge version selection,
it's useful to allow external configuration of the CartridgeRepository to support
defining which version of a cartridge is the "latest" to facilitate migration
testing.

Add a JSON configuration file cartridge_version_overrides.json to the repository
which is a hash of cartridge names to versions. If this file is present, use the
version specified in the file (if present) for a given cartridge name rather than
the version specified in the parameters to #select.