Bug 975794: Use install to create volatile environment variables
Variables which are volatile (in that they may be modified later by the cartridge and
are expected to persist between version upgrades) shouldn't be created via the `env`
system in the cartridge unless they are ERB based; otherwise, the variables will overwrite
already created files in instantiated cartridges each time an upgrade occurs.

Create these variables in the `install` script so that they are primed during new cartridge
installation and not touched during subsequent version upgrades.