Bug 980102
The original link $OPENSHIFT_REPO_DIR/../node_modules is wrong.
The one we need is $OPENSHIFT_NODEJS_DIR/node_modules, which
will be used in build() if deplist.txt is present.

There is another occurrence of 'npm install', but that happens in
$OPENSHIFT_REPO_DIR, to which we have write access.

For existing nodejs applications, we simply remove the dangling symlink.

To test, create a Node.js app, then add a module (say 'coffee-script')
to deplist.txt, add and commit this change, and push.