Bug 974983
Set up $PATH so that ~/.gem/bin precedes /opt/rh/ruby193/root/usr/bin or
/usr/bin (as appropriate).

This effectively reverts fcc9026, which added RubyGems' path during the
user setup. That commit was incomplete, since that led to inconsistency
with the 1.9 apps (see
https://bugzilla.redhat.com/show_bug.cgi?id=974983#c8).

Instead, we use $OPENSHIFT_RUBY_PATH_ELEMENT to ensure that $PATH is
correct. (It is possible to use $OPENSHIFT_RUBYGEMS_PATH_ELEMENT still,
but that relies on $*_PATH_ELEMENT processing order, which we need to
avoid here.)