Bug 984811
As is the case with a very old cartridge, if both
OPENSHIFT_POSTGRES_VERSION and OPENSHIFT_POSTGRESQL_VERSION do
not exist in postgresql/env directory, create one with '8.4'.

Hard coding this value should be OK, since 8.4 was the only supported
version of PostgreSQL for the cartridge this condition ever applies.