Bug 989467
This commit does *not* fix the problem described in the Bugzilla.
Rather, when git clone fails when using SSH, suggest using a different
URL scheme so that the authentication is less of a problem.

Here, the SSH URL detection is purely formal.
The passed URL is deemed SSH if one of the following condition is true:
  1. it starts with 'ssh:'
  2. it contains '@' and not '//' (consider http://user:pass@example.com/foo.git)