<admin/stats> add unit tests, fix a bug
Bug was that all nodes would be considered inactive
because it was not requesting the district_active fact.

Removed model/admin coverage exclusion. There were some
pieces of code that were too hard to exercise, will be left
for integration tests that can create users and apps for
the purpose.