Bug 994922
Warn user of supervisor control when we detect 'hot_deploy' marker.

When the hot_deploy marker is present, stop nor start is invoked, so
we need to check the presence of the marker in a different way.

Notice that the first time we push 'hot_deploy' marker, the stop cycle
will not warn that it detected the marker (because it doesn't exist on
the gear).