<Admin::Stats> serializable results, :read_file option
Previously it seemed like making results serializable was a concern for
the thing serializing them. But then, that requires the consumer of the
results to know about the need for doing that. The default hash procs
that make results unserializable are only needed for building results,
so it seemed like a better idea to just remove those in the results
returned. So that's what we do now.

Admin::Stats can be instructed to load the results from a JSON or YAML
file rather than generate them from the DB and MCollective.

Ensure openshift/data_store is loaded - did not always seem to be.