plugins/dns/nsupdate: Specify zone
Make the nsupdate DNS plug-in send the zone, if it is set, to the DNS
server.  The nsupdate plug-in inherits the BIND_ZONE setting from the
bind plug-in on which it was based, but it was not actually sending the
zone even when it had it.

This commit fixes bug 1000184.