<admin suggestions> allow 1-node districts
Previously districts always got at least 2 nodes (else why have them?)
however in the case where gears are just used for compartmentalized
storage and almost all are idle, it can make sense to have just one node
in a district. The district then doesn't serve much purpose, but you'll
need it if you have set districts to be required for the installation.
It's sort of an odd use case but it is allowed.