<oo-diagnostics> Improve regexes for tests on Origin
The mcollective SELinux context check and the Apache certificate check
both failed in Origin on Fedora 19. This was due to regexes which
weren't robust enough to account for the use of ruby-mri to run
mcollective in Fedora and the inversion of the vhost output from
Fedora's "httpd -S". This commit updates the regexes to account for
these conditions while maintaining existing compatibility with
RHEL-based hosts.