Allow numeric-only Rails app names
In Bug 970150, we introduced the check to enforce an arbitrary
restriction that the Rails application names cannot start with a
numeral.

While it is true that `rails new` fails if the application name starts
with a numeral, this restriction should not prevent OpenShift from
hosting such an application.

Our quickstart now has a workaround:
https://github.com/openshift/rails-example/pull/29

The issue should be addressed in `mysql2` gem as well.