<capacity suggestions> bug 1004686
https://bugzilla.redhat.com/show_bug.cgi?id=1004686
Testers noted that the fill algorithm for putting new nodes in districts
was kind of weird. It was attempting to keep from filling districts
prematurely, but this was the wrong approach. What we do now is default
to a low active percentage setting that will result in a fairly small
district size target. Then, we suggest filling existing districts before
creating new districts as full as needed.