oo-diagnostics: Fix check for mlocate package
test_altered_package_owned_configs: Check the exit code of `rpm -q mlocate`
instead of checking the output.

The old code erroneously assumed that the rpm command would give no output
if the package were not installed, and so it would always return true even
if mlocate were not present.