Don't use app_dns == gear_dns with update_cluster
Previously update_cluster was only syncing content to and activating
new gears if the proxy gear in question was the original one
(OPENSHIFT_APP_DNS == OPENSHIFT_GEAR_DNS). Instead of doing this,
modify the broker to select a proxy and pass a boolean indicating that
proxy should do the syncing and activating.