Bug 1024698
Fix update_cluster to handle the case where a cartridge's minimum
scaling count is >= 2 - treat all web gears except self as new in this
situation.

Don't rotate out when stopping a gear if it's called during the initial
build from post-configure.

Don't check for the force_clean_build marker in prereceive if it's
called during the initial build from post-configure; default to false
instead.

Don't distribute or attempt to do remote activations during an initial
build (this will be taken care of when update-cluster is invoked
afterward).