Gear groups should timeout aggressively when fetching state
The gear group state call needs to time out more quickly on absent gears.  Set default to 10s, and allow clients to request a specific timeout via the API.  Verified that partial results are returned.  UI will use 3 second timeout, and show 'unknown' for any that have not returned yet.

Also, exclude quota checking from app-state-show because it is not intended to be read by clients and because it is a performance critical path