Bug 1026367 - Improving oo-admin-ctl-iptables-port-proxy "UI"
Previously the restart, reload, stop and status commands were noop.  This can
be confusing for admins.

Now:

stop: flushes the rhc-app-comm chain
restart: calls stop then start
reload: calls start
status: removed

The motivation for this is the following:

1 Admin sees something weird with the rhc-app-comm chain not working as expected
2 Admin runs `for s in 'iptables network openshift-iptables-port-proxy'; do server $s restart'
3 Admin notices rhc-app-comm is still empty and is really confused

In this case the event that trigged #1 is that net.ipv4.conf.all.route_localnet
wasn't set in /etc/sysctl.conf.