<oo-accept-node> clarify uid errmsg when non-gear user is in the range
The error message was for processes owned by UIDs in the gear UID range.
However it stated the UID belonged to a gear no longer on the system. In
my case, they actually belonged to a non-gear user with UID 1000.
Technically that's a problem of sorts (could cause problem w/ districts)
so I still want the error, but tried to make it clearer.