Add depends on bind-utils for oo-diagnostics
oo-diagnostics uses the host command, which is in the bind-utils package,
so make rubygem-openshift-origin-common, which includes the oo-diagnostics
command, depend on the bind-utils package.

The above can be a problem on an OpenShift node host, where there may be no
other package installed that pulls in the bind-utils package.  On an
OpenShift broker host, the rubygem-openshift-origin-dns-nsupdate and
openshift-origin-broker-util packages will pull in the bind-utils package.

To be extra safe, add a check in prereq_domain_resolves to warn and skip
the test if the host command cannot be found.

This commit fixes bug 1058527.