Allow new HA app to be created successfully
update_cluster previously assumed that there was always an initial proxy
gear, and that when adding a new proxy gear, it would always be an
increase from e.g. 1 to 2 or more.

Fix update_cluster so it will work correctly when a brand new app is
created with ha=true, in which case the initial call to update_cluster
will contain 2 proxy entries