oo-accept-broker: improve err handling & more
https://trello.com/c/CBLIqG00/214-1-oo-diagnostics-improvements

Somewhere along the way, oo-accept-broker stopped handling broker load
and run errors gracefully. Error output got combined with stdout, and
then that was truncated to the expected number of lines and used as the
expected response, meaning we got bits of stack trace instead of the
configuration values intended, and tried to actually use those.

Now, any error output is routed and noted separately. The error bubbles
up properly and is generally only noted once.

Also, a lot of separate calls invoking the broker for settings were
combined into a single call. This makes the script execute in about half
the time it required before.

Also, some settings were potentially being read from the production
settings file when the development one should have been used. This is
fixed too.

The order of the tests was changed around a little bit but everything
should otherwise work the same.