Bug 980562 - jgroups initial_hosts ordering causing 30s blocking for non-primary gears
The timeout is 30s waiting for number_initial_hosts to respond (==1).
For all gears the order is the same (coming from the hooks). It should
be ordered that the 1st gear in the list is always the current gear vs
always being the primary gear 1st. This results in all of the
non-primary gears to block for 30s while waiting for the primary gear to
come up.