oo-accept-broker: handle broker conf errors better
bug 1078117
https://bugzilla.redhat.com/show_bug.cgi?id=1078117

Solve a couple problems.
First, when invoking the broker to get conf values, handle blank
values correctly. They were not being included, resulting in mismatch
compared to values requested.
Second, should number of values actually do mismatch for some reason,
modify error so it is evident that it's a bug, not the admin's problem.