Switch my.cnf.erb to use innodb_file_per_table.
This prevents users' tablespaces from growing and using the whole gear
quota. Without this option, users will eventually get to a point where
they are out of space, and they must back up, remove the MySQL
cartridge, and then re-add it and restore, which I doubt is the smooth
PaaS experience we want to give them.