Redirect streams to /dev/null unless verbose enabled
Redirect stdout and stderr to /dev/null unless the verbose option
is specified. This is a usability enhancement for the majority
use case which is a daemon-like mode for logshifter in which no
output should ever be expected from the program. If the streams
are left open and undirected, readers may block waiting for
output from logshifter itself.

Resolves bug https://bugzilla.redhat.com/show_bug.cgi?id=1085068