Restorer: Use dbus-send, not oddjob_request
oo-restorer-wrapper.sh: Use dbus-send instead of oddjob_request, and
specify a higher timeout.

oddjob_request has a fixed timeout of 25 seconds, which means that
oo-restorer-wrapper.sh was timing out and returning before the gear was
finished unidling.  As a consequence, restorer.php was sending the HTTP 302
redirect before the gear was ready, so the client would follow the redirect
and get a HTTP 503.

This commit fixes bug 1085489.