Add the ability to adjust when haproxy shutsdown the app cart in the lead gear
https://bugzilla.redhat.com/show_bug.cgi?id=1013166

Currently, haproxy, when scaled to 3 or more gears, will shutdown the
application cartridge (i.e. ruby, python, php, etc) in the gear that
also houses haproxy so that haproxy has access to the entire gear's
resources. It is requested that the number of gears required to shutdown
the application cartridge be adjustable.

The goal of this is to allow the application creator to have one haproxy
gear and just one application gear or , if load permits, have 3 total
gears and still have the application cartridge running on the haproxy
gear.