Speed up chcon calls
Fix matchpathcon_worker thread so it truly waits when it has no requests
to process, instead of accidentally running a busy loop similar to a
spin lock. It previously was only checking to see if the @context Hash
used for matchpathcon requests and replies was empty, but it's possible
that it's not empty but all the requests have been handled. In this
case, the worker thread is really just waiting for all the callers to
pick up their replies, but instead of waiting to be signaled, it was
looping over and over without broadcasting anything.