Downloadable cartridges are improperly described as getting updates
https://bugzilla.redhat.com/show_bug.cgi?id=1068797

The application types page tags imported downloadable cartridges with
the cartridge icon.  This icon is associated with the following text:
"A managed runtime that receives security updates and upgrades
automatically."

However downloadable cartridges do not receive updates or automatic
upgrades.  This text should be changed, since we already have the shield
icon to indicate cartridges which receive updates.