rhcsh: Fix quota check if quotas disabled
When testing whether the gear's quota usage is close to its quota limit,
substitute 0 for the empty string in case there was an error reading the
current usage.

This changes fixes a problem that shows up, for example,  if quotas are
disabled:

    > quota
    Disk quotas for user [uuid] (uid [uid]): none

Previously, in this situation, the following error message would appear
when the application developer tried to log into the gear using ssh:

    awk: END { exit ( !(  > 90.0)); }
    awk:                  ^ syntax error
    awk: END { exit ( !(  > 90.0)); }
    awk:                         ^ syntax error
    awk: END { exit ( !(  > 90.0)); }
    awk:                          ^ syntax error
    awk: END { exit ( !(  > 90.0)); }
    awk:                  ^ syntax error
    awk: END { exit ( !(  > 90.0)); }
    awk:                         ^ syntax error
    awk: END { exit ( !(  > 90.0)); }
    awk:                          ^ syntax error

This commit fixes bug 1092722.